registerObject wrapper around registerFN
Right now you have pass each individual function in an object to the
registerFN function to make it profile. This wrapper allows users to
pass entire objects, with each of their functions getting added to the
profiler.

This will affect both direction functions and those assigned through
the prototype attribute, allowing complex functions that work with the
“new” keyword to easily be added to the profiler.